some super mario world textures
note - these textures are borrowed under fair use
they are not intended to be part of any release game, and are
being used entirely for placeholder/educational purposes
the only exception should be if I decide to actually write some
kind of mario parody...or to simply ignore laws that in their
current state are rather absurd.

second note - the preceeding is probably simply a sign I'm a bit
less than entirely sane at the moment.  we shall see how the coommit
log fares from here on in ^^;;